+++ packages/react/CHANGELOG.md
@@ -11,6 +11,7 @@
 * GreyVest: Give nested tables a grey background
 * GreyVest ResultPager: hide disabled items, add margin between left/right page and +/-5 page, use chevron icons 
 * GreyVest: Update tab padding
+* GreyVest: Facet spacing tweaks
 
 # 1.25.0
 * Made the checkable tables capable of selecting all the visible results.